backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlmodel==0.0.19
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.9
pytest==8.2.2
httpx==0.27.0
alembic==1.13.1
pydantic-settings==2.3.4
requests==2.32.5
tabulate==0.9.0
